% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_data_apple.R
\docType{data}
\name{apple}
\alias{apple}
\title{Apple stock price}
\format{
A zoo series with 2347 data:
\describe{
\item{price}{stock close price, in US dollars}
}
}
\source{
\url{https://www.quandl.com/}
}
\usage{
apple
}
\description{
A dataset containing the daily close stock prices of Apple company from
2020-01-01 to 2018-12-31.
}
\keyword{datasets}
