% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_apply_ptsl.R
\name{fl_apply_ptsl}
\alias{fl_apply_ptsl}
\title{Apply ptsl on t1}
\usage{
fl_apply_ptsl(
  x,
  events,
  lower_barrier = TRUE,
  upper_barrier = TRUE,
  lower_multiplier = 1,
  upper_multiplier = 1
)
}
\arguments{
\item{x}{A time series data}

\item{events}{A data.frame, with t1 (the time stamp of vertical barrier) and
trgt (the unit width of horizontal barriers)}

\item{lower_barrier}{A Boolean, indicating if applying lower horizontal barrier}

\item{upper_barrier}{A Boolean, indicating if applying upper horizontal barrier}

\item{lower_multiplier}{An integer, the factor that multiplies trgt to set
the width of the lower barrier}

\item{upper_multiplier}{An integer, the factor that multiplies trgt to set
the width of the upper barrier}
}
\value{
A data frame, the first touch of three barriers
}
\description{
This function calculates the first time the time series data touch the
triple barriers, two horizontal barriers and one vertical barrier. Users can
set barriers width.
}
\examples{
x <- apple
fl_apply_ptsl(x, fl_simulate_events(x))

}
\author{
Yi Mi
}
