% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_cusum_filter.R
\name{fl_cusum_filter}
\alias{fl_cusum_filter}
\title{Apply symmetric CUSUM filter}
\usage{
fl_cusum_filter(x, h, just_dates = F)
}
\arguments{
\item{x}{A zoo time series data we with to filter}

\item{h}{An integer of the threshold (the filter size)}

\item{just_dates}{A Boolean, indicating if returning simply a vector of dates
or a zoo time series data}
}
\value{
A vector of dates or a zoo time series data
}
\description{
This function apply symmetric CUSUM filter, which is a quality-control
method, designed to detect a shift in the mean value of a measured quantity
away from a detect value.
}
\examples{
fl_cusum_filter(apple, 10)

}
\author{
Yi Mi
}
