% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_daily_volatility.R
\name{fl_daily_volatility}
\alias{fl_daily_volatility}
\title{Compute an asset's daily price volatility}
\usage{
fl_daily_volatility(x, span = 100)
}
\arguments{
\item{x}{A \code{zoo} time series object}

\item{span}{A numeric vector of days}
}
\value{
A \code{zoo} time series object of daily price volatilities
}
\description{
This function computes the daily price volatility at intraday estimation
points, applying a span of days to an exponentially weighted moving standard
deviation.
}
\examples{
fl_daily_volatility(apple, 100)

}
