% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_get_daily_volatility.R
\name{fl_get_daily_volatility}
\alias{fl_get_daily_volatility}
\title{Compute daily volatility}
\usage{
fl_get_daily_volatility(x, span = 100)
}
\arguments{
\item{x}{A time series data we want to compute daily volatility}

\item{span}{An integer of span days}
}
\value{
A time series data of daily volatility
}
\description{
This function computes daily volatility at intraday estimation points,
applying a span of \code{span} days to an exponentially weighted moving standard
deviation. We can use the output of this function to set default profit
taking and stop-loss limits.
}
\examples{
x <- apple
fl_get_daily_volatility(x, 100)

}
\author{
Yi Mi
}
