% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_get_index.R
\name{fl_get_index}
\alias{fl_get_index}
\title{Get time index of observations started and ended}
\usage{
fl_get_index(label_df)
}
\arguments{
\item{label_df}{A data.frame with at least two columns,
"first_touch": Time when the observation ended (i.e. when the
barriers are touched),
"label": An integer label for the observation,
and the index of label_df should be the time when the observation started}
}
\value{
A data.frame of time index and label
}
\description{
This function compute the time index of observations started and ended (when
the barriers are touched), used for k-fold CV and creating data partition.
}
\examples{
label_df <- fl_get_label(apple, fl_simulate_events(apple))
fl_get_index(label_df)

}
\author{
Yi Mi
}
