% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_get_label.R
\name{fl_get_label}
\alias{fl_get_label}
\title{Triple-barrier labeling method}
\usage{
fl_get_label(x, events, vertial_touch_label = "sign")
}
\arguments{
\item{x}{A zoo time series data}

\item{events}{A data.frame of events with \code{t1} and \code{trgt}, can be simulated
by calling function simulate_events
"t1": a series of time stamp of the vertical barrier
"trgt": the unit absolute return used to set up horizontal barrier.}

\item{vertial_touch_label}{An argument with two options "sign" and "zero"
"sign" (the label is decided by the sign of the return)
"zero" (if vertical touch first then set label to 0)}
}
\value{
A data.frame, combining the events, first touch and label
}
\description{
This function finds the time of the first barrier touched.
}
\examples{
fl_get_label(apple, fl_simulate_events(apple))

}
\author{
Yi Mi
}
