% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_partition_data.R
\name{fl_partition_data}
\alias{fl_partition_data}
\title{Create data partition}
\usage{
fl_partition_data(df, purged = TRUE, p = 0.8)
}
\arguments{
\item{df}{A data.frame with at least three columns,
"t0_index": Time index for the observation started,
"first_touch_index": Time index for the observation ended,
i.e. when the barriers are touched
"label": Integer label for the observation}

\item{purged}{A Boolean indicating whether using purging and embargo or not}

\item{p}{The percentage of training set}
}
\value{
If purged is true, return a list of k-fold training, test and purged sets,
If purged is false, return a list of k-fold training and test sets
}
\description{
This function divides the data into a training set and a test set,
the first 80\% is the training set, and the last 20\% is the test set.
}
\examples{
label_df <- fl_get_label(apple, fl_simulate_events(apple))
df <- fl_label_index(label_df)
fl_partition_data(df)

}
\author{
Yi Mi
}
