% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_plot_cusum.R
\name{fl_plot_cusum}
\alias{fl_plot_cusum}
\title{Visualize the symmetric CUSUM filter result}
\usage{
fl_plot_cusum(x, cusum)
}
\arguments{
\item{x}{A \code{zoo} time series object of daily stock prices.}

\item{cusum}{A filtered zoo time series object using CUSUM filtering.}
}
\value{
A \code{plot} object
}
\description{
This function displays the symmetric CUSUM filter result by plotting the
asset's price and points that identify the filtered data.
}
\examples{
cusum <- fl_cusum_filter(apple, 10)
fl_plot_cusum(apple, cusum)

}
