% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_plot_filter_result.R
\name{fl_plot_filter_result}
\alias{fl_plot_filter_result}
\title{Plot symmetric CUSUM filter result}
\usage{
fl_plot_filter_result(x, filtered_x)
}
\arguments{
\item{x}{A time series data of daily stock price in zoo format}

\item{filtered_x}{Filtered x using CUSUM}
}
\value{
The function prints the plot
}
\description{
This function plots symmetric CUSUM filter result by plotting the line of
stock price with dots of filtered data.
}
\examples{
x <- apple
filtered_x <- fl_filter_symmetric_cusum(x, 10)
fl_plot_filter_result(x, filtered_x)

}
\author{
Yi Mi
}
