% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_purge_CV.R
\name{fl_purge_cv}
\alias{fl_purge_cv}
\title{Partition data for k-fold cross-validation with purging and embargo}
\usage{
fl_purge_cv(df, purged = TRUE, k = 5)
}
\arguments{
\item{df}{A data frame with at least three columns.
\code{t0_index}: the time index for the start observations.
\code{first_touch_index}: the time index for the end observation
when the barriers are touched.
\code{label}: a numeric label for the observation.}

\item{purged}{A logical vector indicating whether to purge observations or
not.}

\item{k}{A numeric vector indicating the number of folds.}
}
\value{
If \code{purged} is \code{TRUE}, return a list of k-fold training,
test and purged sets. If \code{purged} is \code{FALSE}, return a list of
k-fold training and test sets.
}
\description{
This function partitions time series data observations for k-fold
cross-validation with purging and embargo. Observations from the training
set are purged to prevent leakage from the training set to testing set.
Embargo is also used to reduce serial correlation.
}
\examples{
label_df <- fl_get_label(apple, fl_simulate_events(apple))
df <- fl_label_index(label_df)
fl_purge_cv(df)

}
