% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_simulate_events.R
\name{fl_simulate_events}
\alias{fl_simulate_events}
\title{Simulate events}
\usage{
fl_simulate_events(x, delta = 15, span = 100)
}
\arguments{
\item{x}{A \code{zoo} time series object.}

\item{delta}{The time interval between day 0 and t1.}

\item{span}{Span days.}
}
\value{
A data frame of events.
\code{t1}: the time stamp of vertical barriers.
\code{trgt}: the unit absolute return used to set up horizontal
barrier.
}
\description{
This function simulates events to get the time stamp of vertical barriers
and the unit absolute return which is used to set horizontal barriers as
unit width. We get the simulate \code{trgt} using the function
\code{fl_daily_volatility()}.
}
\examples{
fl_simulate_events(apple)

}
