% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finma-package.R
\name{globalVariables}
\alias{globalVariables}
\title{The global variables used for this package.

Inside get_index function, their are two variable used as global,
"first_touch_index" and "t0_index"}
\description{
The global variables used for this package.

Inside get_index function, their are two variable used as global,
"first_touch_index" and "t0_index"
}
\references{
Prado, M. L. (2018). Advances in financial machine learning. New Jersey: Wiley.
}
\author{
Yi Mi
}
