% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_data_apple.R
\docType{data}
\name{apple}
\alias{apple}
\title{Apple (AAPL) stock price}
\format{
A \code{zoo} series with 2,347 time points

\describe{
\item{price}{closing stock price in US dollars}
}
}
\source{
\url{https://www.quandl.com/}
}
\usage{
apple
}
\description{
A time-series of the daily closing stock price of Apple (AAPL) from
2010-01-01 to 2018-12-31.
}
\keyword{datasets}
