% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_cusum_filter.R
\name{fl_cusum_filter}
\alias{fl_cusum_filter}
\title{Apply a symmetric CUSUM filter}
\usage{
fl_cusum_filter(x, h, just_dates = FALSE)
}
\arguments{
\item{x}{A \code{zoo} time series object to filter.}

\item{h}{A numeric vector of length one for the threshold (the filter size).}

\item{just_dates}{A logical vector of length one. A value of \code{TRUE}
returns a vector of dates, otherwise a \code{zoo} series is returned.}
}
\value{
A vector of dates or a \code{zoo} time series object.
}
\description{
This function applies a symmetric CUSUM filter, which is a quality-control
method, designed to detect a shift in the mean value of a measured quantity
away from a target value.
}
\examples{
fl_cusum_filter(apple, 10)

}
