% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_get_label.R
\name{fl_get_label}
\alias{fl_get_label}
\title{Triple-barrier labeling method}
\usage{
fl_get_label(x, events, vertial_touch_label = "sign")
}
\arguments{
\item{x}{A \code{zoo} time series object}

\item{events}{A data frame of events, can be simulated by calling function
simulate_events.
\code{t1}: a series of time stamp of the vertical barrier.
\code{trgt}: the unit absolute return used to set up horizontal
barrier.}

\item{vertial_touch_label}{An argument with two options: "sign" and "zero".
For "sign", the label is decided by the sign of the return. For "zero",
if vertical barrier is touched first, then the label is set to 0.}
}
\value{
A data frame of events with the first barrier touch label
}
\description{
This function identifies which of the three barriers (
horizontal-upper, horizontal-lower, vertical) is touched first and gives the
label of -1, 0, or 1.
}
\examples{
fl_get_label(apple, fl_simulate_events(apple))

}
