% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_partition_data.R
\name{fl_partition_data}
\alias{fl_partition_data}
\title{Partition time series data for training and testing}
\usage{
fl_partition_data(df, purged = TRUE, p = 0.8)
}
\arguments{
\item{df}{A data frame with at least three columns.
\code{t0_index}: the time index for the start observations.
\code{first_touch_index}: the time index for the end observation
when the barriers are touched.
\code{label}: a numeric label for the observation.}

\item{purged}{A logical vector indicating whether to purge observations or
not.}

\item{p}{The percentage of observations allocated to the training set.}
}
\value{
If \code{purged} is \code{TRUE}, return a list of k-fold training,
test and purged sets. If \code{purged} is \code{FALSE}, return a list of
k-fold training and test sets
}
\description{
This function divides the data into train and test sets. The first 80\% is
observations are assigned to the training set; the last 20\% are assigned to
the testing set.
}
\examples{
label_df <- fl_get_label(apple, fl_simulate_events(apple))
df <- fl_label_index(label_df)
fl_partition_data(df)

}
