% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_plot_cusum.R
\name{fl_plot_cusum}
\alias{fl_plot_cusum}
\title{Plot symmetric CUSUM filter result}
\usage{
fl_plot_cusum(x, cusum)
}
\arguments{
\item{x}{A zoo time series data of daily stock price in zoo format}

\item{cusum}{Filtered x using CUSUM}
}
\value{
The function prints the plot
}
\description{
This function plots symmetric CUSUM filter result by plotting the line of
stock price with dots of filtered data.
}
\examples{
cusum <- fl_cusum_filter(apple, 10)
fl_plot_cusum(apple, cusum)

}
\author{
Yi Mi
}
