% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_purge_cv.R
\name{fl_purge_cv}
\alias{fl_purge_cv}
\title{k-fold CV with purging and embargo}
\usage{
fl_purge_cv(df, purged = TRUE, k = 5)
}
\arguments{
\item{df}{A data.frame with at least three columns,
"t0_index": Time index for the observation started,
"first_touch_index": Time index for the observation ended,
i.e. when the barriers are touched
"label": Integer label for the observation}

\item{purged}{A Boolean indicating whether using purging and embargo or not}

\item{k}{An integer for k-fold}
}
\value{
If purged is true, return a list of k-fold training, test and purged set,
If purged is false, return a list of k-fold training and test set
}
\description{
This function implements purged k-fold CV to purge from training set all
observations whose labels overlapped in time with those labels included in
the testing set to reduce leakage, which is "purging". In addition, it also
eliminates from the training set observations that immediately follow an
observation in the testing set to reduce serial correlation, which is
"embargo".
}
\examples{
label_df <- fl_get_label(apple, fl_simulate_events(apple))
df <- fl_label_index(label_df)
fl_purge_cv(df)

}
\author{
Yi Mi
}
