% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_simulate_events.R
\name{fl_simulate_events}
\alias{fl_simulate_events}
\title{Simulate events}
\usage{
fl_simulate_events(x, delta_t1 = 15, span = 100)
}
\arguments{
\item{x}{A time series data we want to simulate}

\item{delta_t1}{The interval between day 0 and t1, which is 15 as default}

\item{span}{Span days}
}
\value{
A data frame of events with t1 and trgt
"t1": The time stamp of vertical barriers
"trgt": the unit absolute return used to set up horizontal barrier.
}
\description{
This function simulates events to get the time stamp of vertical barriers
and the unit absolute return which is used to set up horizontal barrier as
unit width. We get the simulate \code{trgt} using the function
fl_get_daily_volatility().
}
\examples{
x <- apple
fl_simulate_events(x)

}
\author{
Yi Mi
}
