% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_touch_barriers.R
\name{fl_touch_barriers}
\alias{fl_touch_barriers}
\title{Get the time of the first barrier touch}
\usage{
fl_touch_barriers(
  x,
  events,
  lower_barrier = TRUE,
  upper_barrier = TRUE,
  lower_multiplier = 1,
  upper_multiplier = 1
)
}
\arguments{
\item{x}{A \code{zoo} time series object.}

\item{events}{A data frame, with \code{t1} - the time stamp of vertical
barrier, and \code{trgt} - the unit width of horizontal barriers.}

\item{lower_barrier}{A logical vector to apply a lower horizontal
barrier or not.}

\item{upper_barrier}{A logical vector to apply a upper horizontal
barrier or not.}

\item{lower_multiplier}{A multiplicative numeric quantity to set the width
of the lower barrier.}

\item{upper_multiplier}{A multiplicative numeric quantity to set the width
of the upper barrier.}
}
\value{
A data frame with the first touch of three barriers.
}
\description{
This function identifies the time when the first of three barriers (
horizontal-upper, horizontal-lower, vertical) is touched.
}
\examples{
fl_touch_barriers(apple, fl_simulate_events(apple))

}
